minimized Wordle.js and exported logic to GameManager.js
