+++ src/Wordle/styles/wordle.css
@@ -60,3 +60,12 @@ h1.wordle-header {
     flex-wrap: wrap;
     margin: 0 auto;
 }
+
+::-webkit-scrollbar {
+    display: none;
+}
+
+.instruction code {
+    background-color: #363636;
+    padding: 2px 3px 1px 2px;
+}
